reorganized project structure into feature-based folders
