+++ src/app/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import authReducer from '../features/auth/authSlice';
 import { authApi } from '@/features/auth/authApi';
 import newsReducer from '@/features/news/newsSlice';
-import mediaReducer from '@/features/newsadd/mediaSlice';
+import mediaReducer from '@/features/newsadd/media/mediaSlice';
 
 export const store = configureStore({
   reducer: {
