+++ src/components/newsadd/MediaPicker.tsx
@@ -1,6 +1,6 @@
 import { useDispatch } from 'react-redux';
 import { MediaMockData } from './MediaMockData';
-import { addImage, closePicker } from '@/features/newsadd/mediaSlice';
+import { addImage, closePicker } from '@/features/newsadd/media/mediaSlice';
 
 export default function MediaPicker() {
   const dispatch = useDispatch();
